# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel API Documentation",
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/notes": {
            "get": {
                "tags": [
                    "Notes"
                ],
                "summary": "Получить все заметки пользователя",
                "operationId": "0eb12475438792d91075dbcb643f0184",
                "responses": {
                    "200": {
                        "description": "Список заметок"
                    },
                    "204": {
                        "description": "У вас нет заметок"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Notes"
                ],
                "summary": "Создать новую заметку",
                "operationId": "b156674541dd281cfa5e1eef5b81e762",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "header": {
                                        "type": "string",
                                        "example": "Новая заметка"
                                    },
                                    "text_note": {
                                        "type": "string",
                                        "example": "Содержание заметки"
                                    },
                                    "tags": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "tag_name": {
                                                    "type": "string",
                                                    "example": "Тест"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Заметка создана",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "header": {
                                            "type": "string"
                                        },
                                        "text_note": {
                                            "type": "string"
                                        },
                                        "tags": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "tag_name": {
                                                        "type": "string"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/notes/{id}": {
            "get": {
                "tags": [
                    "Notes"
                ],
                "summary": "Получить заметку по ID",
                "operationId": "573e9e975b61116618fc65ac194f73a2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Заметка найдена",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "header": {
                                            "type": "string"
                                        },
                                        "text_onote": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Заметка не найдена"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Notes"
                ],
                "summary": "Обновить заметку",
                "operationId": "5b5a63bd38624e4dd88279410049b69d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "header": {
                                        "type": "string",
                                        "example": "Обновленный заголовок"
                                    },
                                    "text_note": {
                                        "type": "string",
                                        "example": "Обновленное содержание"
                                    },
                                    "tags": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "tag_name": {
                                                    "type": "string",
                                                    "example": "Тег"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Заметка обновлена"
                    },
                    "404": {
                        "description": "Заметка не найдена"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Notes"
                ],
                "summary": "Удаление заметок по ID",
                "operationId": "c1580766728fabf7cd42a700ecb6a3ef",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the note",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Заметка удалена успешно"
                    },
                    "404": {
                        "description": "Заметка не найдена"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Notes"
                ],
                "summary": "Обновить заметку",
                "operationId": "36878c9583285b08a2be17d5bd65b312",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "header": {
                                        "type": "string",
                                        "example": "Обновленный заголовок"
                                    },
                                    "text_note": {
                                        "type": "string",
                                        "example": "Обновленное содержание"
                                    },
                                    "tags": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "tag_name": {
                                                    "type": "string",
                                                    "example": "Тег"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Заметка обновлена"
                    },
                    "404": {
                        "description": "Заметка не найдена"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/notes/search": {
            "post": {
                "tags": [
                    "Notes"
                ],
                "summary": "Поиск заметок по тегам",
                "operationId": "3a5117462e60ab0fe39a4d936299c648",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "tags": {
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "example": "Тест"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Заметки найдены",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "header": {
                                            "type": "string"
                                        },
                                        "text_note": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "204": {
                        "description": "Результата нет",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "header": {
                                            "type": "string"
                                        },
                                        "text_note": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/tags": {
            "get": {
                "tags": [
                    "Tags"
                ],
                "summary": "Получение списка тегов",
                "description": "Возвращает список всех тегов",
                "operationId": "e320cb59dba3e323dacbbc502509717c",
                "responses": {
                    "200": {
                        "description": "Список тегов",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Tag"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Теги не найдены"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Tags"
                ],
                "summary": "Создание нового тега",
                "operationId": "d5b48558ae6b21f20e758584734e2cd7",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "tag_name": {
                                        "type": "string",
                                        "example": "Имя тега"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Тег создан успешно"
                    },
                    "400": {
                        "description": "Ошибка валидации"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Tag": {
                "title": "Tag",
                "description": "Модель тега",
                "required": [
                    "id",
                    "tag_name"
                ],
                "properties": {
                    "id": {
                        "description": "Идентификатор тега",
                        "type": "integer"
                    },
                    "tag_name": {
                        "description": "Название тега",
                        "type": "string"
                    },
                    "user_id": {
                        "description": "ID пользователя, которому принадлежит тег",
                        "type": "integer"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}